# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/orders": {
      "get": {
        "operationId": "OrdersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "หน้าที่ต้องการ เริ่มที่ 1",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "จำนวนต่อหน้า 1–100",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "orderStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "CONFIRMED",
                "PREPARING",
                "READY_FOR_PICKUP",
                "SHIPPED",
                "COMPLETED",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "paymentStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "WAITING_VERIFY",
                "PAID",
                "REJECTED",
                "REFUNDED"
              ]
            }
          },
          {
            "name": "deliveryMethod",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PICKUP",
                "DELIVERY"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "ค้นจากเลขที่คำสั่งซื้อ รหัสรับสินค้า เลขพัสดุ ชื่อหรืออีเมลผู้สั่ง",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "orderNumber",
            "required": false,
            "in": "query",
            "description": "เลขที่คำสั่งซื้อแบบตรงตัว เช่น ORD-202609-0001",
            "schema": {
              "maxLength": 20,
              "type": "string"
            }
          },
          {
            "name": "mine",
            "required": false,
            "in": "query",
            "description": "true = เฉพาะคำสั่งซื้อของฉัน (เจ้าหน้าที่ใช้ดูของตัวเอง)",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/OrderModel"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "รายการคำสั่งซื้อ — ลูกค้าเห็นเฉพาะของตัวเอง เจ้าหน้าที่เห็นทั้งหมด",
        "tags": [
          "orders"
        ]
      },
      "post": {
        "operationId": "OrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/OrderModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "สั่งซื้อ — เจ้าของคือผู้ที่ login อยู่",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/OrderModel"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "orders"
        ]
      },
      "delete": {
        "operationId": "OrdersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeletedModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ลบคำสั่งซื้อ (ผู้ดูแล)",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/v1/orders/{id}/payment-status": {
      "patch": {
        "operationId": "OrdersController_updatePaymentStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/OrderModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ตรวจสลิป / เปลี่ยนสถานะการชำระเงิน (เจ้าหน้าที่)",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/v1/orders/{id}/order-status": {
      "patch": {
        "operationId": "OrdersController_updateOrderStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/OrderModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "เปลี่ยนสถานะคำสั่งซื้อ (เจ้าหน้าที่) — ยกเลิกแล้วคืนสต็อก",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/api/v1/orders/{id}/payment-slip": {
      "patch": {
        "operationId": "OrdersController_updatePaymentSlip",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentSlipDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/OrderModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "แนบสลิปการชำระเงิน — เจ้าของคำสั่งซื้อ",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/v1/orders/{id}/items": {
      "get": {
        "operationId": "OrdersController_findItems",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/OrderItemModel"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "รายการสินค้าในคำสั่งซื้อ",
        "tags": [
          "orders"
        ]
      },
      "post": {
        "operationId": "OrdersController_addItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/OrderItemModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "เพิ่มสินค้าเข้าคำสั่งซื้อ (เจ้าหน้าที่)",
        "tags": [
          "orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "CSMJU Shop API",
    "description": "ร้านค้าออนไลน์สาขาวิทยาการคอมพิวเตอร์ ม.แม่โจ้ — ทุก endpoint ต้องมี Core Hub token (Authorization: Bearer หรือคุกกี้ core_hub_access_token) ยกเว้น GET /api/health · ทุก response ห่อด้วย { success, data, meta? } · เงินทุกช่องเป็นจำนวนเต็มหน่วยสตางค์",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PageMetaModel": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
...
      "DeletedModel": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "deleted",
          "id"
        ]
      },
...
      "DeliveryMethod": {
        "type": "string",
        "enum": [
          "PICKUP",
          "DELIVERY"
        ]
      },
      "PaymentStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "WAITING_VERIFY",
          "PAID",
          "REJECTED",
          "REFUNDED"
        ]
      },
      "OrderStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "CONFIRMED",
          "PREPARING",
          "READY_FOR_PICKUP",
          "SHIPPED",
          "COMPLETED",
          "CANCELLED"
        ]
      },
      "OrderItemProductModel": {
        "type": "object",
        "properties": {
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "isPreorder": {
            "type": "boolean"
          }
        },
        "required": [
          "imageUrl",
          "id",
          "name",
          "category",
          "isPreorder"
        ]
      },
      "OrderItemVariantModel": {
        "type": "object",
        "properties": {
          "product": {
            "$ref": "#/components/schemas/OrderItemProductModel"
          },
          "id": {
            "type": "string"
          },
          "variantName": {
            "type": "string"
          }
        },
        "required": [
          "product",
          "id",
          "variantName"
        ]
      },
      "OrderItemModel": {
        "type": "object",
        "properties": {
          "variant": {
            "$ref": "#/components/schemas/OrderItemVariantModel"
          },
          "id": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "variantId": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "unitPrice": {
            "type": "number",
            "description": "ราคาต่อชิ้น ณ ตอนสั่ง หน่วยสตางค์"
          },
          "isPreorderItem": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "variant",
          "id",
          "orderId",
          "variantId",
          "quantity",
          "unitPrice",
          "isPreorderItem",
          "createdAt",
          "updatedAt"
        ]
      },
      "PickupLogEntryModel": {
        "type": "object",
        "properties": {
          "notes": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "staffCoreUserId": {
            "type": "string"
          },
          "staffEmail": {
            "type": "string"
          },
          "pickupTime": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "notes",
          "id",
          "orderId",
          "staffCoreUserId",
          "staffEmail",
          "pickupTime",
          "createdAt",
          "updatedAt"
        ]
      },
      "OrderModel": {
        "type": "object",
        "properties": {
          "deliveryMethod": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DeliveryMethod"
              }
            ]
          },
          "shippingAddress": {
            "type": "string",
            "nullable": true
          },
          "paymentSlip": {
            "type": "string",
            "nullable": true
          },
          "paymentStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentStatus"
              }
            ]
          },
          "orderStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderStatus"
              }
            ]
          },
          "pickupCode": {
            "type": "string",
            "nullable": true,
            "description": "รหัสรับสินค้า (เฉพาะ PICKUP)"
          },
          "shippingCarrier": {
            "type": "string",
            "nullable": true
          },
          "trackingNumber": {
            "type": "string",
            "nullable": true
          },
          "shippedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "orderItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemModel"
            }
          },
          "pickupLogs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PickupLogEntryModel"
            }
          },
          "id": {
            "type": "string"
          },
          "orderNumber": {
            "type": "string",
            "description": "เช่น ORD-202609-0001"
          },
          "coreUserId": {
            "type": "string",
            "description": "เจ้าของคำสั่งซื้อ (claim `sub` ของ Core Hub)"
          },
          "customerEmail": {
            "type": "string"
          },
          "recipientName": {
            "type": "string"
          },
          "recipientPhone": {
            "type": "string"
          },
          "totalAmount": {
            "type": "number",
            "description": "ยอดรวม หน่วยสตางค์"
          },
          "shippingFee": {
            "type": "number",
            "description": "ค่าจัดส่ง หน่วยสตางค์"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "deliveryMethod",
          "shippingAddress",
          "paymentSlip",
          "paymentStatus",
          "orderStatus",
          "pickupCode",
          "shippingCarrier",
          "trackingNumber",
          "shippedAt",
          "orderItems",
          "pickupLogs",
          "id",
          "orderNumber",
          "coreUserId",
          "customerEmail",
          "recipientName",
          "recipientPhone",
          "totalAmount",
          "shippingFee",
          "createdAt",
          "updatedAt"
        ]
      },
      "OrderItemInputDto": {
        "type": "object",
        "properties": {
          "variantId": {
            "type": "string",
            "description": "id ของตัวเลือกสินค้า",
            "format": "uuid"
          },
          "quantity": {
            "type": "number",
            "description": "จำนวน",
            "minimum": 1,
            "maximum": 99
          }
        },
        "required": [
          "variantId",
          "quantity"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "deliveryMethod": {
            "type": "string",
            "description": "PICKUP = รับที่สาขา · DELIVERY = จัดส่ง",
            "enum": [
              "PICKUP",
              "DELIVERY"
            ]
          },
          "recipientName": {
            "type": "string",
            "description": "ชื่อผู้รับ",
            "minLength": 1,
            "maxLength": 100
          },
          "recipientPhone": {
            "type": "string",
            "description": "เบอร์โทร ตัวเลขล้วน 9–10 หลัก ไม่มีขีด",
            "pattern": "^0\\d{8,9}$"
          },
          "shippingAddress": {
            "type": "string",
            "description": "ที่อยู่จัดส่ง — บังคับเมื่อ deliveryMethod = DELIVERY",
            "minLength": 0,
            "maxLength": 500
          },
          "paymentSlip": {
            "type": "string",
            "description": "URL หรือชื่อไฟล์สลิปโอนเงิน (แนบภายหลังได้)",
            "minLength": 1,
            "maxLength": 255
          },
          "items": {
            "description": "รายการสินค้า",
            "minItems": 1,
            "maxItems": 50,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemInputDto"
            }
          }
        },
        "required": [
          "deliveryMethod",
          "recipientName",
          "recipientPhone",
          "items"
        ]
      },
      "UpdatePaymentStatusDto": {
        "type": "object",
        "properties": {
          "paymentStatus": {
            "type": "string",
            "description": "สถานะการชำระเงินใหม่",
            "enum": [
              "PENDING",
              "WAITING_VERIFY",
              "PAID",
              "REJECTED",
              "REFUNDED"
            ]
          }
        },
        "required": [
          "paymentStatus"
        ]
      },
      "UpdateOrderStatusDto": {
        "type": "object",
        "properties": {
          "orderStatus": {
            "type": "string",
            "description": "สถานะคำสั่งซื้อใหม่",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "PREPARING",
              "READY_FOR_PICKUP",
              "SHIPPED",
              "COMPLETED",
              "CANCELLED"
            ]
          }
        },
        "required": [
          "orderStatus"
        ]
      },
...
      "UpdatePaymentSlipDto": {
        "type": "object",
        "properties": {
          "paymentSlip": {
            "type": "string",
            "description": "URL หรือชื่อไฟล์สลิปโอนเงิน",
            "minLength": 1,
            "maxLength": 255
          }
        },
        "required": [
          "paymentSlip"
        ]
      },
      "CreateOrderItemDto": {
        "type": "object",
        "properties": {
          "variantId": {
            "type": "string",
            "description": "id ของตัวเลือกสินค้า",
            "format": "uuid"
          },
          "quantity": {
            "type": "number",
            "description": "จำนวน",
            "minimum": 1,
            "maximum": 99
          }
        },
        "required": [
          "variantId",
          "quantity"
        ]
      },
...
    }
  }
}